+++ packages/core/CHANGELOG.md
@@ -30,6 +30,7 @@
 * Improve performance of `deserializeSerializationChunk` function ever so slightly.
 * Remove the unused (and non-performant) `NaiveSymbolTable` class.
 * Improve performance of `nodeSerializer` function a tiny bit.
+    * Fix a bug where some features’ values are serialized multiple times.
 * Expose feature resolution, through the `featureResolversFor` function (and associated types), which performs proper checking.
   This avoids undebuggable `undefined` dereferencing at runtime in `deltaDeserializer` and `eventToDeltaTranslator` functions.